// src/lib/types/auth.d.ts
export type LoginResponse = {
    "token": string,
    "user": {
        _id: string,
        username: string,
        username: string,
        firstName: string,
        lastName: string,
        email: string,
        phone: string,
        role: string,
        isVerified: boolean,
        createdAt: string
    }
}